This index file constists of html and css styling.
This is not responsive. This is work is one of project from Frontend masters course Complete Intro to Web Development, v2 by brian holt. This index.html just to represent my work in css and html(I think this is not productive code or best code but this can be improved). I hope i will improve and make code more productive and responsive